Support Serializing Joi .array().items() Schemas
Because this module relies on the Joi schema in order to determine how to serialize a value, values in nested arrays are not serialized properly.

This fixes the issue by properly storing the DynamoDB data type for schema-validated Joi arrays, and also ensures that the array items are serialized recursively.